Minor improvement on egg
easier to derive classes from BaseObject

+++ tests/run_all_tests.py
@@ -1,5 +1,6 @@
 import unittest as _ut
-import shutil as _sh
+import shutil   as _sh
+import os       as _os
 
 from test__databox   import *
 from test__fitter    import *
@@ -9,6 +10,6 @@ from test__dialogs   import *
 from test__egg       import *
 from test__edu       import *
 
-_sh.rmtree('gui_settings')
+if _os.path.exists('gui_settings'): _sh.rmtree('gui_settings')
 
 _ut.main()
